fix(user-repository): rename folder of user domain repositories
