+++ src/User/Entities/User.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\User\Entities;
 
-use App\User\Repository\UserRepository;
+use App\User\Repositories\UserRepository;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Security\Core\User\PasswordAuthenticatedUserInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
